+++ src/components/CardFoldout.tsx
